split_1/2d header rearrangement, a few std:: straglers
