+++ src/diagnostic/detail/h5writer.h
@@ -318,7 +318,7 @@ void Writer<ModelView>::initializeDatasets_(std::vector<DiagnosticProperties*> c
 
     // sets empty vectors in case current process lacks patch on a level
     std::size_t maxMPILevel = core::mpi::max(maxLocalLevel);
-    for (size_t lvl = maxLocalLevel; lvl <= maxMPILevel; lvl++)
+    for (std::size_t lvl = maxLocalLevel; lvl <= maxMPILevel; lvl++)
         if (!lvlPatchIDs.count(lvl))
             lvlPatchIDs.emplace(lvl, std::vector<std::string>());
 
@@ -351,7 +351,7 @@ void Writer<ModelView>::writeDatasets_(std::vector<DiagnosticProperties*> const&
     modelView_.visitHierarchy(writePatch, minLevel, maxLevel);
 
     std::size_t maxMPILevel = core::mpi::max(maxLocalLevel);
-    for (size_t lvl = maxLocalLevel; lvl <= maxMPILevel; lvl++)
+    for (std::size_t lvl = maxLocalLevel; lvl <= maxMPILevel; lvl++)
         if (!patchAttributes.count(lvl))
             patchAttributes.emplace(lvl, std::vector<std::pair<std::string, Attributes>>{});
