+++ src/diagnostic/detail/types/electromag.h
@@ -1,7 +1,10 @@
 #ifndef PHARE_DIAGNOSTIC_DETAIL_TYPES_ELECTROMAG_H
 #define PHARE_DIAGNOSTIC_DETAIL_TYPES_ELECTROMAG_H
 
+#include "diagnostic/detail/h5file.h"
 #include "diagnostic/detail/h5typewriter.h"
+
+#include "core/data/vecfield/vecfield_component.h"
 
 namespace PHARE::diagnostic::h5
 {
