+++ src/diagnostic/detail/types/fluid.h
@@ -1,7 +1,10 @@
 #ifndef PHARE_DIAGNOSTIC_DETAIL_TYPES_FLUID_H
 #define PHARE_DIAGNOSTIC_DETAIL_TYPES_FLUID_H
 
+#include "diagnostic/detail/h5file.h"
 #include "diagnostic/detail/h5typewriter.h"
+
+#include "core/data/vecfield/vecfield_component.h"
 
 namespace PHARE::diagnostic::h5
 {
